CWB v3.4.11: re-factored handling of compressed files; now centrally supported by cl_open_stream() and cl_close_stream()
git-svn-id: https://svn.code.sf.net/p/cwb/code/cwb/trunk@929 b72dbb9a-5b02-45d8-a000-b69eecba7de2

+++ cl/makecomps.c
@@ -267,7 +267,7 @@ creat_rev_corpus(Component *revcorp)
   assert(freqs->corpus == revcorp->corpus); /* gotta be kidding ... */
 
   lexsize = cl_max_id(attr);        /* this is the number of lexicon entries for this attribute */
-  ptab = (int **) cl_malloc(sizeof(int *) * lexsize); /* table of pointers into <buffer> */
+  ptab = (int **) cl_malloc(sizeof(int *) * ((size_t) lexsize)); /* table of pointers into <buffer> */
 
   /* determine REVCORP size (== number of tokens) */
   datasize = cl_max_cpos(attr);
